feat: move grpc hostname fetching behind serverless tag
